refactor: Rename file to better capture its functionality
